split code, add styles for 1280
move buttons to the separate component , refactoring, update for 1280pc
